+++ NEWS.md
@@ -1,4 +1,5 @@
 # master
 
+* raise custom error classes instead of Exception [anujbiyani]
 * Store unit name
 * Fixed issue with binary prefix parsing in regards to JEDEC